+++ action_server/docs/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## 0.0.28 - 2024-03-11
 
+- `pydantic` models are accepted as the input and output of `@action`s.
+
 - The action package name is now gotten from the `package.yaml` and not from the directory name
   (it's still gotten from the directory name when `conda.yaml` is used for backward compatibility).