+++ action_server/docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Support parsing Custom Types in Action Server UI action run view
+- Add Console output to Action Server UI action run view
 
 ## 0.0.28 - 2024-03-11